requirements.txt
APScheduler==3.6.3
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
cachetools==4.2.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Django==4.1.4
exceptiongroup==1.0.4
fake-useragent==1.1.1
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-resources==5.10.1
libretranslatepy==2.1.1
lxml==4.9.2
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tornado==6.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
whitenoise==6.2.0
wsproto==1.2.0
zipp==3.11.0
